Restructure base.py and basic.py.
Now we have properties.py, primitive.py, basic.py, nested.py, hasproperties.py
