+++ properties/math.py
@@ -1,4 +1,4 @@
-"""math.py: vectormath property classes"""
+"""math.py: array and vectormath property classes that require numpy"""
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
@@ -8,7 +8,7 @@ import numpy as np
 from six import integer_types, string_types
 import vectormath as vmath
 
-from .basic import Property
+from .properties import Property
 
 TYPE_MAPPINGS = {
     int: 'i',
